requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
altair>=5.2.0
python-dotenv==1.0.1
duckdb>=1.1.0
duckdb-engine>=0.13.0
sqlalchemy>=2.0.30
llama-index-core>=0.10.40
llama-index-llms-gemini
llama-index-embeddings-gemini
google-generativeai>=0.7.0
pycountry>=24.6.1